clinical data
Consider using builder pattern: If there are too many fields to set, it's better to use the builder pattern to simplify object creation.
Used final fields
Used primitive types, it's better to use primitive types instead of wrapper classes to save memory.
Add validation that the fields cannot be null. However, this does not validate the format or content of the fields. It's better to add validation to ensure that the fields have valid values.